requirements.txt.txt
streamlit
pandas
numpy
matplotlib
scikit-learn
nltk
summa
pymupdf
pillow
python-docx
requests
beautifulsoup4
